+++ packages/Webkul/Product/src/Jobs/Search/IndexProducts.php
@@ -1,48 +1,36 @@
 <?php
 
-namespace Webkul\Product\Jobs\ElasticSearch;
+namespace Webkul\Product\Jobs\Search;
 
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Queue\SerializesModels;
-use Webkul\Product\Helpers\Indexers\ElasticSearch;
 use Webkul\Product\Repositories\ProductRepository;
+use Webkul\Product\Services\Search\SearchEngineManager;
 
-class UpdateCreateIndex implements ShouldQueue
+class IndexProducts implements ShouldQueue
 {
     use Dispatchable, InteractsWithQueue, Queueable, SerializesModels;
 
     /**
      * Create a new job instance.
-     *
-     * @param  array  $productIds
-     * @return void
      */
-    public function __construct(protected $productIds)
-    {
-        $this->productIds = $productIds;
-    }
+    public function __construct(protected array $productIds) {}
 
     /**
      * Execute the job.
-     *
-     * @return void
      */
-    public function handle()
+    public function handle(SearchEngineManager $manager, ProductRepository $productRepository): void
     {
-        if (core()->getConfigData('catalog.products.search.engine') != 'elastic') {
-            return;
-        }
-
         $ids = implode(',', $this->productIds);
 
-        $products = app(ProductRepository::class)
+        $products = $productRepository
             ->whereIn('id', $this->productIds)
             ->orderByRaw("FIELD(id, $ids)")
             ->get();
 
-        app(ElasticSearch::class)->reindexRows($products);
+        $manager->indexer()->indexBatch($products->all());
     }
 }