+++ packages/Webkul/Shop/src/Http/Controllers/SearchController.php
@@ -4,6 +4,7 @@ namespace Webkul\Shop\Http\Controllers;
 
 use Illuminate\View\View;
 use Webkul\Marketing\Repositories\SearchTermRepository;
+use Webkul\Product\Enums\SearchContextEnum;
 use Webkul\Product\Repositories\SearchRepository;
 
 class SearchController extends Controller
@@ -47,12 +48,8 @@ class SearchController extends Controller
             ! request()->has('suggest')
             || request()->query('suggest') !== '0'
         ) {
-            $searchEngine = core()->getConfigData('catalog.products.search.engine') === 'elastic'
-                ? core()->getConfigData('catalog.products.search.storefront_mode')
-                : 'database';
-
             $suggestion = $this->searchRepository
-                ->setSearchEngine($searchEngine)
+                ->setSearchContext(SearchContextEnum::STOREFRONT)
                 ->getSuggestions($query);
         }